web/requirements.txt
flask
Flask-WTF
flask-sqlalchemy
flask-security
mariadb
reportlab
xhtml2pdf